Version Packages (#55)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## azdo-npm-auth@1.8.2

### Patch Changes

- 5a87531: It's not obvious we are publishing with [trusted
publishing](https://docs.npmjs.com/trusted-publishers) - this release
removes the npm token; this way if we succeed in publishing it must mean
we are using trusted publishing.

I'm not certain if changesets is calling `npm publish` under the bonnet,
this should hopefully determine if it is.

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>